Better refactoring of 3D code base
